Extract shared dynlib error types into internal/dynlib package
Move ErrRecursionDepthExceeded, ErrLibraryHashMismatch, ErrEmptyLibraryPath,
and ErrDynLibDepsRequired from elfdynlib (and ErrRecursionDepthExceeded from
machodylib) into a new format-independent internal/dynlib package.

ErrRecursionDepthExceeded was duplicated verbatim in both packages.
ErrDynLibDepsRequired was already returned for both ELF and Mach-O paths in
verification.Manager. ErrLibraryHashMismatch and ErrEmptyLibraryPath are used
by elfdynlib.DynLibVerifier which verifies deps for both binary formats.
