+++ .golangci.yml
@@ -45,6 +45,7 @@ linters:
             - $gostd
             - github.com/isseis/go-safe-cmd-runner/internal/cmdcommon
             - github.com/isseis/go-safe-cmd-runner/internal/common
+            - github.com/isseis/go-safe-cmd-runner/internal/dynlib
             - github.com/isseis/go-safe-cmd-runner/internal/elfdynlib
             - github.com/isseis/go-safe-cmd-runner/internal/fileanalysis
             - github.com/isseis/go-safe-cmd-runner/internal/machodylib
@@ -78,6 +79,7 @@ linters:
             - github.com/joho/godotenv
             - github.com/isseis/go-safe-cmd-runner/internal/cmdcommon
             - github.com/isseis/go-safe-cmd-runner/internal/common
+            - github.com/isseis/go-safe-cmd-runner/internal/dynlib
             - github.com/isseis/go-safe-cmd-runner/internal/elfdynlib
             - github.com/isseis/go-safe-cmd-runner/internal/fileanalysis
             - github.com/isseis/go-safe-cmd-runner/internal/filevalidator
