+++ internal/verification/manager_test.go
@@ -12,7 +12,7 @@ import (
 
 	"github.com/isseis/go-safe-cmd-runner/internal/common"
 	commontesting "github.com/isseis/go-safe-cmd-runner/internal/common/testutil"
-	"github.com/isseis/go-safe-cmd-runner/internal/elfdynlib"
+	"github.com/isseis/go-safe-cmd-runner/internal/dynlib"
 	"github.com/isseis/go-safe-cmd-runner/internal/fileanalysis"
 	"github.com/isseis/go-safe-cmd-runner/internal/filevalidator"
 	"github.com/isseis/go-safe-cmd-runner/internal/runner/runnertypes"
@@ -1554,7 +1554,7 @@ func TestVerify_ELFNoDynLibDeps(t *testing.T) {
 	verifyErr := m.VerifyCommandDynLibDeps(cmdPath)
 	require.Error(t, verifyErr)
 
-	var errRequired *elfdynlib.ErrDynLibDepsRequired
+	var errRequired *dynlib.ErrDynLibDepsRequired
 	assert.ErrorAs(t, verifyErr, &errRequired, "expected ErrDynLibDepsRequired for dynamic ELF without DynLibDeps")
 }
